scripts/{bigpulp,hero}: Rework to fix bugs and remove duplication
The script for standalone bigPULP was outdated and missed fixes that had
been applied to the HERO script.  To fix this and prevent it from
happening again, the bulk of commands are now in a common script that is
sourced during both HERO and standalone bigPULP setup.

Additionally, the standalone bigPULP setup script now only downloads
dependencies if the `PULP_RISCV_GCC_TOOLCHAIN` environment variable is
not set.