This commit adds support for compiling with the PULP LLVM toolchain.
The usage is analogous to the GCC toolchain: users must define the environmental variable "PULP_RISCV_LLVM_TOOLCHAIN".
Only one toolchain can be defined; if both env variables are present, the make system terminates with an error message.

Note that the fixes to the *.S files are general, but the related bugs are only triggered using LLVM.